+++ src/services/get-aggregate-service.js
@@ -6,6 +6,7 @@ import {
   gridCodeLayerStyle
 } from '../utils/mapHelper'
 import { mapHexGridFromJson, mapGridCodeFromJson } from '../utils/searchHelper'
+import { appendStacHeaderCookies } from '../utils/stacRequest'
 
 export async function AggregateSearchService(searchParams, gridType) {
   const requestHeaders = new Headers()
@@ -15,6 +16,8 @@ export async function AggregateSearchService(searchParams, gridType) {
   if (JWT && isSTACTokenAuthEnabled) {
     requestHeaders.append('Authorization', `Bearer ${JWT}`)
   }
+  appendStacHeaderCookies(requestHeaders)
+
   await fetch(
     `${
       store.getState().mainSlice.appConfig.STAC_API_URL